🚚 Move string arbitraries into /arbitrary
