+++ src/check/arbitrary/ObjectArbitrary.ts
@@ -11,7 +11,8 @@ import { maxSafeInteger } from '../../arbitrary/maxSafeInteger';
 import { memo, Memo } from '../../arbitrary/memo';
 import { oneof } from '../../arbitrary/oneof';
 import { set } from '../../arbitrary/set';
-import { string, unicodeString } from './StringArbitrary';
+import { string } from '../../arbitrary/string';
+import { unicodeString } from '../../arbitrary/unicodeString';
 import { tuple } from '../../arbitrary/tuple';
 import { bigInt } from '../../arbitrary/bigInt';
 import { date } from './DateArbitrary';
